feat: took out useClickOutside to Hooks
